(2/352 pass) Adjust tests for great victory
The two unit tests failed because of the passthrough types. Since the
`x` passed into R function was used in the python function, its type was
not known. I'm still not entirely happy, since the nexus usage
statements show the unknown types when they should be known. This is a
fairly minor quibble that can be fixed in the future. For now I won't
worry about it. I can clean up the typechecker implementation when I
refactor with recursion schemes. That will be an epic refactoring for
sure.

The module-form-110 test failed because the sourced `sum` function was
not given a concrete type. In theory, the type could be inferred by
transferring over the type from the declaration. But this isn't entirely
obvious and violates the hard rule that all sources should have concrete
types assigned, rather than inherit laterally. Explicit is better than
implicit.