Fix bug when import name != module name
Previously, if you import a module like so, `import foo`, and then you
find a file in the right place (e.g., `foo.loc`), but the module name is
wrong (e.g., `module bar`), an infinite loop occured. The loop occurred
because `parseImports` takes a list of expected imports and recursively
parses them, storing found module names in a set. Since the expected
name is never seen, the function recurses forever.

I solved this by passing the expected module name to the parser wrapped
in Just. If the morloc code is root (a script or string), then this is
Nothing and the module name, if present, will be un-constratained. If we
are following an import link, the module name will be matched in the
parser, raising a syntax error on failure, e.g.,:

```
SyntaxError: /home/z/.morloc/src/bio/io/main.loc:1:8:
  |
1 | module yolo.bio.io
  |        ^^^^^^
unexpected "yolo.b"
expecting "bio.io"
```