Begin dev cycle 0.38.0
One annoying but necessary decision I need to make is about casing. I do
not want to have to flip-flop in the future about snakecase versus camel
case and the handling of acronyms. Here is a motivating example, what
name should we use for a function that writes a PNG file?

 1. `write_png_file`
 2. `write_PNG_file`
 3. `writePNGFile`
 4. `writePngFile`
 5. `WritePNGFile`

(1) is readable. But PNG is an acronym, so maybe (2) is more
appropriate, though it is a bit ugly. (3) is the conventional camel
case, but the "GF" capitals next to eachother looks weird. One way to
avoid adjacent capitals is to treat acronyms as words (4). (5) shows
Pascal case, but I am reserving this for types.

Types tend to have shorter names. They are typically nouns, perhaps with
an adjective attached. Functions, in contrast, tend to be verbs, often
with objects. Pascal case, with all words capitalized, then, works fine
for types. I will keep the Haskell convention for types.

For individual terms, I do like snakecase, but when functions are
applied, things get ugly. Compare these two alternatives:

```
json_list_filter_map parse_record select_record gisaid_file
jsonListFilterMap parseRecord selectRecord gisaidFile
```

In the first application, it is hard to even tell what is being applied.
THe underscores and the spaces are too similar. In the second line, it
is quite clear that there are three arguments. Camel-case makes broader
chunks of text more readable by making *phrase* parsing easier.
Snake-case makes terms more readable by improving term parsing. Though
this difference in term parsing is slight, lower than the difference in
phrase parsing.

But how common are these long phrases of long composite terms? Usually
there are subexpressions, so the expression may be:

```
json_list_filter_map (parse_record arg1 42) (select_record "asdf") (gisaid_file True)
jsonListFilterMap (parseRecord arg1 42) (selectRecord "asdf") (gisaidFile True)
```

When the text is broken up by more arguments and parentheses, the
parseability of the two conventions is more similar. Though still the
camel-case seems clearer to my eyes.

Parsing may also be improved with formatting, such as vertical
alignment:

```
json_list_filter_map
    parse_record
    select_record
    gisaid_file
```

Another consideration is cognitive load. For lower-snake-case, where
terms are not allowed to have capitals, there will be fewer reasonable
variants for each name. Specifically, camel-case has two sub-conventions
for dealing with acronyms. These can either be treated as words or not.
Treated as words, only the first letter of the acronym is upper-case.
This approach avoids adjacent acronyms (except for 1-letter acronyms)
and avoids the case of names turning into Pascal case if they start with
a capital. For example, the term name `PDFFile`, would be `pdfFile`. So
word-camel-case seems better to me. Though it is not an easy convention
to make syntactically required.

But usually these long names come from writing entire sentences into the
function name. These sentences, with subjects, verbs, adjectives, and
objects are playing the roles of adhoc type signatures and mnemonic
devices. Look at this signature:

```
convert_pdf_to_png :: PDF -> PNG
```

There is no new information in the `pdf_to_png` phrase. We could write:

```
convert :: PDF -> PNG
```

Functions should be as generic as possible, and generic necessarily
means short. And the types carry the subject and object of the verb. So
perhaps function names should just be the verb, whenever possible. Names
should not be a place to store documentation. They should be terse
identifiers with mnemonic value, but not so long that they dominate the
phrases. So the value of term parsing is lower than the value of phrase
parsing.

Cases are boring. They cause unnecessary cognitive load (one more thing
to remember and fight over). So I want to force the case convention such
that breaking it causes a syntax error. I can't enforce camel case
(there is no way for the compiler to know which letters to capitalize).
But I can enforce snake case. I can force term variable names to be all
lowercase, '_', and maybe primes.

I can lay out the pros and cons for term types:

 * term parsing - +1 snake
 * phrase parsing - +3 camel
 * cognitive load - +3 snake - more enforceable
 * uniformity - +2 camel - no mix of camel and snake
 * brevity - +1 camel - no extra chars
 * compatibility - +2 camel - matches Haskell and everything on Hoogle

Camel-case seems to win for terms. For types, I'm already going with
Haskell conventions. That leaves modules and packages. Packages must be
lower-case, since some operating systems don't distinguish case. Modules
currently are lower case, but why exactly?

What about modules? Haskell uses Pascal case, so does the rest of the ML
family. Python uses lower case names (by convention, not lexing rule)
and optional, but discouraged, underscores. Module names will need '.'
to separate namespaces. The module names also need to map to file names
on the file system.

Pros and cons for module names:
 * conservative - +1 pascal - matches conventions in ML languages
 * file naming -  +1 snake - matches more common file naming conventions
   (and decreases SHIFTing)

I need to rethink my package system. Currently there is a one-to-one
relationship between packages and modules set in the `morloc install`
command, where the first field in the module name is the github username
(defaulting to morloc-lib) and the second is the repo name. But packages
should be able to contain many modules.

Another consideration is that "morloc" packages in the future may be
fully functional other-language packages with just a single file
specifying morloc types for export. In this case, the directory
structure will be determined by the main language. In this case, there
may be multiple modules *in one file*, the `<pkgname>.mlc` file.
Further, the package name may be constrained as well by the main
language. So the mapping between file and module name cannot be safely
assumed.

If module names do not link to directory or repository names, then they
are free to be whatever we please. This is probably for the best. The
Morloc general language must not be constrained by any specific
language.

Forced conventions:
 * term names are camel case
 * types are camel case (usually fully lower) or pascal (non-generic) case
 * module names - Pascal?
 * package names - mostly free form, preferred lowercase

Yeah, so Haskell all the way?

Here is a useful discussion of Haskell conventions: https://kowainik.github.io/posts/naming-conventions.