Better error messages for typechecking of terms with no expressions
This was occurring becuase my typechecker is trying to typecheck
exported types. For example,

```
type Cpp Real = "double"

export Real
```

This is a reasonable thing to export, since other modules may want to
share type conventions. But the typechecker and code generators
currently assume that everything that is exported is a function. I need
to fix this.