Separate type and term import/exports
Both types and terms may be imported and exported. But the meaning of
these two actions are very different. Previously both types and terms
were stored as `EVar` objects. This was confusing since types are not
expression variables. However, these imported/exported types cannot be
stored as `TVar` either, since they do not yet have laguages associated
with them.

When handling type aliases, we need to be careful that we only follow
type imports. When generating SAnno trees, we need to be sure we only
follow term imports. At the lexing level, we also need to be certain
these two are not confused. To that end I added separate lexers for
upper and lower case names, one types and one for terms. This enforces
the conventions and allows separation into terms at types in the parser.
They are no stored in the `AliasedSymbol name alias` and `Symbol name`
types.