Fix eta reduction
To allow eta-reduction I have to add eta-expansion to the frontend
typechecker. These expansions alter the tree in three locations.

 * synthE `-->E`
 * synthE `-->I==>`
 * checkE (LamS [] _) (FunU (t:_)

At least all eta reduction is in the frontend and occurs immediately
after the required type information on the called functions is
extracted. It is logical, I guess. But I am still not happy. Modifying
the tree at this stage leads to nodes with undefined indices, which I
default to (-1). This is sloppy. I should probably generate free
indices.

I remember back to the great deep/wide quagmire of 2021. At the time,
for reasons that I did not find at all satisfactory, I settled on a wide
style. Under this style I implemented the functions type, for example,
as `Fun [Type] Type` with a list of recursive types for inputs. The deep
alternative would have been the traditional FL definition of `Fun Type
Type`. The wide style made for easier parsing and generating, but less
elegant reasoning and typechecking.

Now with eta reduction, I am encountering another difficulty with the
wide style. It is ambiguous. It allows a functions to be written as `a
-> a -> a` or `a -> (a -> a)`. Lambdas expressions also suffer from the
possibility of nested expressions.

```
Fun ts (Fun ts' t)  ===  Fun (ts <> ts') t
Lam vs (Lam vs' x)  ===  Lam (vs <> vs') x
```

I don't think the deep style would entirely remove the eta complexities,
but it would certainly simplify the reasoning process. And it would
remove the ridiculous need for "normalizeTypes" calls.

The codebase is teetering under its weight. There are so many dark
corners with poorly designed syntax. Not to mention so many cases that I
have barely scratched, like `const` keywords in C++ and
pass-by-reference versus pass-by-value.

But I will make it work. I'll find a way.