Begin dev cycle 40 -- argh
The embarrasing truth is that I can't really instantiate records in
expressions. If I define a record:

```
record Config = Config {name :: Str, version :: Str}
```

and the use it

```
foo :: Config -> Int -> Str
bar = foo {name = "Bob", version "v2.0"}
bar = foo {name = "Bob", version "v2.0"}
```

This is a common case where I partially apply a configuration to bar.
Except I can't do that. This raises a type error. It shouldn't, since I
know that the record there is a `Config` type and I should be able to
check that the fields match.

The goal of this dev cycle is to rectify this problem. It should only
require an extension to the subtyping algorithm.