(4/370 fail) Allow passthroughs to be serialized
I am narrowing in on the serialization error. It occurs when a generic
function in language A wraps a value in a foreign language.

For example

```
import pybase (id)
import rbase (add)

foo py :: "float" -> "float"
foo x = id (add 1.0 x)

export foo
```

The concrete types are:

```
\ x ->
(
  py@id :: py@*a -> py@*a
   ( R@add :: R@numeric -> R@numeric -> R@numeric
     1.0 :: R@numeric
     x :: R@numeric
   ) :: py@*a
) :: py@float
```

`add` returns an R@numeric type. Since the calling function is in
Python3, the return value is cast as of UnkT type.

The serialization error occurs when the generator attempts to unpack the
output of add that needs to be passed to id. It so happens that `id`
does nothing with its input, so receiving and returning an Unk would be
fine (just return the input JSON string). But morloc has no way of
knowing this. It needs to deserialize the output of add into Python.

This deserialization requires a type annotation, since `id` is generic.
And we provided a signature. In the python type signature, we said
the argument should have type "float" and a "float" should be
returned. So `id` must return a py@float. But id has type `a -> a`, so
its input type must equal its output type. So the type should be:

```
\ x ->
(
  py@id :: py@float -> py@float
   ( R@add :: R@numeric -> R@numeric -> R@numeric
     1.0 :: R@numeric
     x :: R@numeric
   ) :: py@float
) :: py@float
```

Something is off in the typechecker.

But let's say we don't give `foo` a type:

```
import pybase (id)
import rbase (add)

foo x = id (add 1.0 x)

export foo
```

Now the following type should obtain:

```
\ x ->
(
  py@id :: py@*a -> py@*a
   ( R@add :: R@numeric -> R@numeric -> R@numeric
     1.0 :: R@numeric
     x :: R@numeric
   ) :: py@*a
) :: py@*a
```

Now the `py@*a` are all passthrough types. It is the deserialization of
passthroughs that has been crashing the program. But what if we just
legalize that shit? Let them "deserialize" into the serial type. Blobs.
Strings or data blobs that can be rearranged but not evaluated. For `id`
and many other generic recombinators, just working with blobs is what we
want.

So two bugs to fix. A type inference bug and a blob bug. The blob bug
is simple and I've fixed that in this commit. Now the serializer will
not die on "SerialUnknown" types, instead it will happily serialize them
as the default serial type for the given language.

Two bugs left (that I know of):
 1. A argument number bug (discussed in previous commit)
 2. The type inference bug discussed above