(3/370 fail) Fix naming bug
interop-2 was failing simply because I was not generating unique
variables in Generate.hs. The non-unique variables alone weren't a
problem, however, I was using identity of arguments between a manifold
and a child manifold to determine which arguments were new to the child
(introduced by lambda binding) and which were passthrough.

See the translaters `call = case splitArgs args pargs ...` code. I
really should not be handling that logic in the translators. It should
be baked into the ExprM grammar.

3 bugs left:
  1. record-promotion-2 - type error
  2. interop-3 - need to wrap foreign calls in manifolds or lambdas
  3. eta-reduction-7 - vanishing unused arguments

These three bugs all seem to have very different causes. So .. chop chop
lots of work left to do.