(2/271 fail) Fix one case of foreign higher order
The code generation for calls to higher-order functions was completely
wrong. Now in R for unapplied pool calls the generators work. I need to
extend this to partially applied functions (where parent manifold values
may need to be passed) and to C++ and Python (where translateManifold
instances for LamM need to be handled).

Cases to consider where
    g l :: (C_l -> D_l) -> A_l -> B_l
    f m :: C_m -> D_m
    h m :: E_m -> C_m -> D_m
    x :: B
    y :: C

1.  foo x = g f x            -- fully unapplied
2.  foo x = g (h 42.0) x     -- applied to constant
3.  foo x y = g (h y) x      -- applied to argument

Then I need to test all supported languages in the roles of g and f.
For N languages this will only require N tests. For example, with C, R,
and Python:

 C P
 R C
 P R

For the three cases and three languages, then, we will need 9 tests:

 1 C P    2 C P    3 C P
 1 R C    2 R C    3 R C
 1 P R    2 P R    3 P R