(11/380 fail) Fix backend typechecking error
I was failing to apply Gamma to the expressions in check LamS.

This lead to incomplete type inference noted in commit 06a019e.
Downstream it lead to incorrect serialization.

On the generator side, I had not originally implemented the Lambda
generators because lambdas have rather spotty and complicated generation
in some languages. It is easier to just create helper functions and call
them. I should probably implement lambda this way in the generators
(though behavior here may be language specific).