(11/380 fail) Generate delcarations for lambdas
I've added a new record, PoolDocs, to R.hs:translateManifold. It
replaces the 3-tuple with a record and greatly clarifies the code. I
also add a much needed field that stores expressions that should be
written prior to the manifolds in a pool. The lambda declarations will
be up there. Currently this is working for R, I need to extend it to
Python and C++.