(11/380 fail) Extend prior commit to Python generator
The reason why interop-3a-rc passes while interop-3a-pr fails is that
the value passed to the pool is not being serialized. The R serializer
takes the integer value and stringifies it, which happens to work in
this case. The Python serializer, however, raises a type error rather
than attempting to coerce the value. So something is still off.