(10/380 fail) Refactor C++ generator with PoolDocs
The erorr in typeOfExprM is being raised at Cpp.hs:335 in interop3a-cp.
There appears to an application of a serialized function. That should
not happen.

Also, in translateManifold, I left the LamM case as a stub. Once this
stub is filled out and the serialized function bug is fixed, then
interop3a-cp should pass.