(5/380 fail) Allow unused terms in top lambdas
Arguments are passed through the manifolds to the places where they are
called. But if they are never called along a particular branch, the
argument will be removed. This logic is handled in the
`Generate.paramterize` function.

Top-level lambdas are removed in `Generate.express` to expose the
application beneath. However, this was leading to arguments that are
specified by the user but not actually accessed being lost. They appear
in the LamS argument list, but are pruned from AppS.

To fix this issue, I've simply replaced the top-level AppS args with the
LamS args.