(4/380 fail) minor fix - consider a rewrite
There is an ectopic scrap of logic duplicated across my translators in
the `translateManifold` function. This function should take the ExprM
input object and dutifully translate it into the language of interest.
But the ManifoldM case further parses the parent and manifold arguments
to determine if the manifold call should be a function call (e.g.,
`m16(x,y)`, a raw function (e.g., m16), or a lambda (e.g., `lambda y:
m16(x, t)`. This piece of logic would be better implemented in
`Generate.express`.