(2/380 fail) All interop tests pass
The code in Generate.hs:express is kludgy as hell. I need work out a
formal framework for handling the serialization and evalutation locals.
For example, in `f x y`, if is in language L and x and y in language M,
which scope are they evaluated in and how are they serialized? They
system I have now is a mess.

eta-reduction-3 is now failing. It did not used to fail, so something I
changed recently broke it. It seems that the system is losing an
argument somewhere.