(7/391 fail) Partial fix to case 8, new tests
For case 8 which handles partial trans calls, I need to partition the
arguments to the trans call into those that are evaluated on the calling
side and those that are evaluated on the foreign side. I've added
handling for this. It is buggy though. I added a new set of tests
simplified from the old interop-3d.* series.