(22/391 fail) All pass but the interop code
There is more simplification I could do in express, I think. But life is
getting better. Now all the serialization bugs are all in one little
box. And I can put that little box in a bigger box, put that box in a
bigger box, mail it to myself, and SMASH it with a HAMMMMERRRR. Does
anyone read my posts?