(15/391 fail) Add PreArgument
Previously I was passing only the argument variable name and index from
`parameterize` to `express`. This requires a bit of searching to
correctly generate fresh variables at the manifold boundaries.

A cleaner solution is to pass the type, id and varname in one data type
(PreArgument) but without the serialization baggage which will be added
latter.