Resolve "shit output" error message
The error was being raised when a export had no associated bindings. Now
I map the export name to the index. This allows undefined terms to be
exported from a module. Now, that is perhaps a bit sketch. But it is
useful as a "hole". Errors will be raised if either the type of the term
cannot be inferred or, later on, when no source or source signature is
found in the backend.