Change C++ to c++17 standard by default
I am currently hard-coding the C++ version. That is a dreadful idea. The
standard should be inferred from the package descriptions of each module
and otherwise default to, perhaps, c++17?

Oh, and I also need to unify the code in all the preludes and create
better checks that the modules are well-formed. Currently morloc is
lenient. This is fine, since it allows partially written modules to be
tested. Maybe we need a flag for stricter module checking?