Remove unnecessary Int from PackMap
Previously I was passing the number of generic terms in the (un)packer
as part of the key (along with the TVar storing the language and type
name). I used both the number of generics and they name to lookup
packers. However, I want to be able to specialize packers by allowing
some of the generics to be concrete.