Restructure library to avoid name conflicts
The morloc `string` module conflicted with the python `string` module.
The python code told the interpreter to search $MORLOCHOME for modules.
This folder contained the `string/string.py` python module that needed
to be loaded. The morloc python string module was loaded as
`string.string`. The first `string` specified to look in folder
"string". The second was the name of the python file "string.py".

However, the Python interpreter first searches for core modules. It
found a module named "string". But then it raised an error when it found
no sub-module named "string" (so "string.string" was not found).

A rather uncouth solution to this is to place all the morloc modules
under a name unlikely to be used outside morloc. So rather than placing
all the modules in "$MORLOCHOME/src", I will place them in a deeper
subfolder.  So "$MORLOCHOME/src/morloc". This means the python string
module will now have the python path "morloc.plain.morloclib.string.string".

So why all those folders?

 1. "morloc", represents a whole morloc context. In the future, it might
    be replaced with morloc versions or such.

 2. "plain" is one of three folders representing different sources of
    modules. The others being "local" and "github". The latter would
    contain specific modules that are not part of a larger system. The
    idea of a "plain" is a little like a Haskell Prelude, but larger.
    The different plains may have different focuses. Some may be more
    specialized for performance. Others for correctness. Or they may
    follow alternative organizational principles.

 3. "morloclib" is the current default (and only) plain. It will
    certainly be killed in the future, which is why I have not bothered
    to give it a better name.