(1/392 fail) Added a field to ExistU that stores record accesses
When a term like `x@name` appears, the typechecker will first try to
synthesize a type for `x`. This will return an existential (`ExistU`)
since nothing is known about `x`. Then the term `x@name` is checked. But
`name` is not a key in the existential. So an error was raised. To deal
with this, I've added the option to store key accesses in existentials.
There is still something wrong in here, but the idea is to store partial
information in the existential that can be subtyped later against the
full record or used to synthesize a full record.