Fix handling of functions with Unit arguments
In functions with forms like:

```
foo :: () -> A
```

Most languages will not pass a literal Null value. So it is better to
compile them to functions with empty argument lists.

Functions like this may often need to be partially applied, for example:

```
foop :: Config -> () -> A
```

Here we are performing some IO operation that yields A and taks a
configuration Config. Partial application yields the function:

```
() -> A
```

Which retains the idea that this is a function touching the system
(something from nothing).

A weirder case is:

```
weird :: () -> A -> B
```

This seems sus, I'm not sure when this would be used. For now, though, I
allow it. All Unit types will be filtered out of the input.