Replace the Null literal term with ()
The "Null" syntax is one more pointless thing to remember. And if we
have "Null" and (), then what is the meaning of an empty tuple? These
are isomorphic concepts (sets of size 1) so they should have the same
syntax.

This syntax also makes the empty tuples impossible to represent, since
they parse as UniE. Further single tuples are impossible to represent
since they reduce to the expression inside them.

Finally, the type-level expression for a unit type is `()`. So it makes
sense that the corresponding term syntax is the same.