Found the reason for an infinite regression bug
All of these cases lead to infinite recursion in the typechecker:

```
-- infinite recursion
g :: [a] -> a
f :: a -> a
f x = g x

-- infinite recursion
g :: (a, b) -> a
f :: a -> a
f x = g x

-- typechecker misses the incompatible types
g :: (a -> b) -> a
f :: a -> a
f x = g x
```

The reason is that an existential type is being expanded into a larger
type that contains the original type. What I need to do check that the
a term does not appear in its expansion.

The third case is an entirely different bug. The typechecker is stopping
evaluation too early and not checking g against f.