Raise an error if a non-existant term is exported
This also fixed the apparent problem of being unable to export packed
types (like Table or Map). The reason these could not be exported is
that they are not explicitly defined anywhere. Maybe they should be.
But there is not currently a way to do that. And I'm not sure what the
point would be anyway.