Add check for self-recursion existential solutions
The reason for the infinite recursions noted in commit c0e61 was the
creation of self-recursive solutions to existential variables. For
example:

    SolvedG: a_q2 = List <a_q2>

These occur when infinite types are inferred, for example, in:

```
g :: [a] -> a
f :: a -> a
f x = g x
```

The solution is to check for the existence of a term within its
definition. If it exists, raise a type error.

Of the three cases mentioned in the prior commit, the first two now are
checked properly, however the last case still incorrectly passes the
type checker:

```
g :: (a -> b) -> a
f :: a -> a
f x = g x
```

The of `f` is inferred as `a -> a`. Apparently the type of `f` is never
checked against `g`.