Remove `import-1` test
This test has linear algebra library dependencies which should not be
required for a basic test.

Retcon - I've changed my mind. This is a test of morloc's ability to use
system libraries. Testing this is important, so it should be retained.
Further, only the GNU math library is required actually, not cblas, and
this is a fairly simple dependency.