Die if no concrete type found for a sourced term
If a sourced term is not explicitly given a concrete type and if no
concrete type can be generaed from a general type, then raise an error.

In the past, type inference would continue. This could often lead to
ambiguous types appearing in the final program with no clear origin.

This commit also solve the earlier bug of having general types role
in unchanged to the concrete type.