Remove self-recursion check for concrete typedefs
There are cases were the defined term may appear on the right. For example:

  type Py (Tree n e l) = "Tree" n e l

Here the general type Tree is mapped to the concrete type "Tree" in Python.
The fact that the general and concrete names are the same is fine. They are
different languages.

Now I do not check for self-recursion if the types are concrete. This is
not the right solution. For example, the type:

  type Py A = A A

Is now legal and it should not be, since it will both cause an infinite
loop and yield contradictory types. At the same time, the following
types are illegal but probably should be legal:

  type (List a) = L a (List a)
  type (Tree n) = Node n [Tree n]