Fix bug in subtype occurs check
This bug appeared in the following code:

```
g :: (a -> b) -> a
f :: a -> a
f x = g x
```

This code is incorrect, since it requires `a` to be the same as
`(a -> b)`. But instead the typecheck happily returned (a->a) as the
type of `f`.

In the verbose typechecking output, this specific line was where things
went South:

```
  : (<ta4>) <: (<ta4> -> <to5>)
```

Replacing the mangled naming, this is `<a>  <:  (<a> -> <b>)`. Where
`<x>` means `x` is existential. This should fail the occur check since
the left hand value `<a>` is among the free variables in the right hand
side. But it did not. The reason was simply that the order of arguments
passed to the occur check function were reversed.

I never found this bug before because I haven't been testing invalid
code.