Identify problem and kick the can
When the morloc script is typechecked:

```
module m (foo)

source Cpp from "foo.hpp" ("g")
source Py from "foo.py" ("g", "f")

g :: a -> a
g Py :: a -> a
g Cpp :: a -> a

f :: Int -> Int
f Py :: "int" -> "int"

foo x = bar (f x)
bar x = g x
```

We get the following:

```
$ morloc typecheck -r ambiguous.loc
foo Python3Lang :: (int@Python3Lang -> Python3Lang@*o_2)
foo ::
```

This leads us to problem #1. This is a bad error message. It exposes
internal names for unknown terms `*o_2` and leaves a hanging `foo ::`
statement.

The reason the Python return type cannot be determined is because the
C++ function `g` is being used. `g` does not know, currently, how to
represent the output of `f` in C++. This type could be inferred, since
`Int` maps to C++ `int`. But this inference is not currently done
(problem #2).

After the C++ g function is run, the result is returned to a python
lambda funtion and that value is returned. So the nexus passes x to
the python pool where `f x` is evaluated and the result is sent to the
C++ g function. The output of g is returned to the python pool. But the
python pool doesn't know how to serialize it. Though again, they type
should be inferrable from general type (problem #2 again).

We can work around #2 by adding a type annotation:

```
foo Py :: "int" -> "int"
foo x = bar (f x)
bar x = g x
```

Typechecking this yields the expected result:

```
$ morloc typecheck -r ambiguous.loc
foo Python3Lang :: (int@Python3Lang -> int@Python3Lang)
foo :: (Int -> Int)
```

But this still fails to compile:

```
$ morloc make ambiguous.loc
morloc: Cannot deserialize passthrough type: Passthrough
```

C++ doesn't know how to deserialize the `Int` type. Problem #2.

But there is another workaround. Refactoring the code yields to the
correct result:

```
foo x = g . x
```

or

```
foo x = bar (f x)
bar = g
```

Both work correctly. The reason is that they do not wrap `g` in a
lambda. Code for them is generated as `g (f x)` rather than
`(\x . g x) (f x)`. In the later case, the lambda term and `g` are
realized separately: the lambda is realized as Python to avoid the
penalty of a cross-language call and `g` is realized as C++. This leads
to problem #3. The realization engine is broken. This is a VERY tricky
algorithm. I need to completely rewrite it and set it on a firm
mathematical footing. This will require several months of hard work. And
I don't want to put in the time at the moment. I can avoid these
problems by avoiding generic polyglot contacts.

I can come back to this problem when I am ready to do a heavy rewrite of
the realization engine and type system.

I need to write this into a github issue and add it to next cycle's
TODO.