Clarify messages for error occuring when a term is not defined
I ought to catch undefined terms earlier. The reason I don't is that it
can be useful to have the typechecker infer the types of these terms
(rather like the use of holes in Haskell). But it becomes a hard to
diagnose liability further down the line. I don't want this message to
be raised in the code generator section.