Thoughts on the future of morloc
The end goal of morloc has always been to allow better reasoning about
code through a higher logical framework. I wanted to write on the
platonic ideal of a language while also not to give up compatibility
with code in all other languages. So I developed a two layer model. The
general type system where we are free to program in function
compositions with all the beauty and elegance of Haskell and the
concrete type system for describing how these ideal forms map to real
languages.

But AI programming is rapidly improving. Soon, if not already, it will
be better at programming than humans. So perhaps we do not need
interoperability with other languages? What we do still need is a way to
describe computation. What we need is language as a tool for thought;
fine control over the programs that are generated; and confidence in the
correctness of the code.

Generative AIs will soon be able to serve as universal compilers and
translators. So I may shift my focus to a new language where the
emphasis is on the ideal description of the computation and the backend
work that has so occupied my time can be delegated to the AI.

Still, before switching languages, I should publish this paper. I've
made progress with morloc and it will be a foundation that future work
will build on. So I need to make the flu case work.