Fix duplicate pool tree bug
When a manifold with a given index and language is called from two
different segments of the bipartite tree, then complete duplicate
subtrees were being generated in the pools. This caused the C++ code to
fail because of duplicate prototypes, duplicate function definitions,
and duplicate entries in the case-switch dispatch statement.

I fixed this by pruning the extra trees in Generate.hs:pool.

There is still a bug with manifolds inside the trees having multiple
identical implementations.