Distinguish between C++ Unit and void
The C++ void type is not allowed in datastructures. For example, the C++
type `tuple<Int, void>` is illegal. This is because `void` represents
an undefined value, rather than a set of 1 element. This is in contrast
to `None` and `NULL` from Python and R, respectively. So I've defined a
`Unit` type as a enum for C++. This type definition is stored in a
library of C++ morloc types that should be installed along with morloc.

This feels a little too opinionated. I would prefer to allow the user to
define this type. Eventually perhaps it can be wrapped in with
`conventions.loc` instead.