Do not duplicate manifolds in C++
The problem was that for a manifold with a given index, there may be
multiple uses within a pool. This led to multiple cases with the same
index in the C++ case switch statements, multiple identical prototypes,
and multiple identical function definitions. All of this is highly
illegal in C++. One alternative would be to distinguish between these
different manifolds. However, that would be rather wasteful and require
more complexity and more code compilation. So instead I keep track of
which manifolds have been generated and ensure everything is made only
once. The code I've written here is a travesty, though. I really should
use a another state monad.