Restrategize flu demo debugging
The morloc side of the flu case study seems to be working fine now. Most
of the problems appear to be on the bio.tree module side. I can debug
bio.tree through the case study, but this is complicated. It required
debugging the generated pool.hpp files, which are very tangled. And it
requires debugging in a larger codebase.

A far saner strategy is to develop a test suite covering the bio.tree
module and ensure all that C++ code is working properly. Once that is
solid, I can return to the case study.