(0/410 fail) fix and test partial manifold bug
If we have a function such as this:

```
foo = snd3 (\a b c -> b) (1, True, "Zanzibar")
```

An error was raised. This was due to a lot of miswiring in express and
reparameterize, but should be all fixed now. Basically, the bound lambda
arguments that were not directly used in the morloc were being ignored.

Now there are just 39 lines of error code in the flu case study. Next,
I'll continue working on tests of the bio.tree module. After the case
study compiles, I'll just be left with the C++ problem of implementing
the alignment and treebuilding algorithms. For now, I'll make very
simplistic solutions that can serve as a baseline for future work.