Add tests for path shadowing bug
In C++, when I import two modules that contain the same source file, the
paths are shadowed. For example, given these morloc modules:

```
module main (*)

import foo (f)
import bar (g)
```

```
module foo (*)

source Cpp from "main.hpp" ("f")
```

```
module bar (*)

source Cpp from "main.hpp" ("g")
```

Then a g++ command such as this will be generated:

```
g++ -I/path/to/bar -I/path/to/baz
```

And the C++ pool will include the two files with:

```
include "main.hpp"
include "main.hpp"
```

This is the cause of the path shadowing problem. It is not a problem in
Python and R, where I simply write the entire pathway. And that is one
solution here, I could remove the `-I` flags from the g++ call and use
absolute paths to the sources. The other option would be to require
packages make the names unique. This could be done by requiring a
particular folder structure such as the following for the module named
"monkey":

```
monkey/
  main.loc
  monkey/
    main.hpp
```

Then in main.loc:

```
source Cpp "monkey/main.hpp"
```

However, this solution forces the morloc programmer to think about
shadowing and adds a constraint to package design. In return, the
generated C++ code that is never intended to be read, which is
immediately compiled to source, is a little tidier. That is a bad trade.
So I should just write the entire path.