Fix bug occuring for functions named "f"
In Python, when a sourced function was named "f", it conflicted with the
function name "f" used in the pull to represent the manifold function
that would be called. I fixed this by mangling the names. A better
solution would be to use a proper namespace for all imported functions.

In fact, I certainly need to import with namespaces since currently I
could source modules that conflict. For example, if I import the two
python modules "foo" and "bar", they will be imported as:

```
from foo import *
from bar import *
```

And there is nothing preventing them from sharing names. Certainly
morloc does not check for this. In fact, the morloc philosophy rather
encourages it. Instead I should use namespaces. Probably the entire
module path.

```
import foo
import bar

foo.myfun(45)
bar.myfun(56)
```

Pretty obvious solution, much better than mangling the freakin' names.

I should do the same in R and C++. This is all a bit of tedious work,
but well worth it.